Add `StypOutput` namespace
