+++ src/producer/style-producer.ts
@@ -2,8 +2,7 @@
  * @packageDocumentation
  * @module @proc7ts/style-producer
  */
-import { NamespaceAliaser, NamespaceDef } from '@proc7ts/namespace-aliaser';
-import { RenderScheduler } from '@proc7ts/render-scheduler';
+import { NamespaceDef } from '@proc7ts/namespace-aliaser';
 import { StypProperties, StypRule } from '../rule';
 import { StypSelector } from '../selector';
 import { StypRenderer } from './renderer';
@@ -83,92 +82,5 @@ export interface StyleProducer {
    * @returns Either appended empty CSS rule, or `target`.
    */
   addRule(selector?: StypSelector.Normalized): CSSRule;
-
-}
-
-/**
- * CSS styles production options.
- *
- * This options are accepted by [[produceStyle]] function.
- *
- * @category Rendering
- */
-export interface StypOptions {
-
-  /**
-   * A document to produce styles for.
-   *
-   * `window.document` by default.
-   */
-  document?: Document;
-
-  /**
-   * Parent DOM node to add stylesheets to.
-   *
-   * `document.head` by default.
-   */
-  parent?: ParentNode;
-
-  /**
-   * A selector to use for root CSS rule.
-   *
-   * `body` by default.
-   *
-   * For custom elements a `:host` selector would be more appropriate.
-   */
-  rootSelector?: StypSelector;
-
-  /**
-   * Creates CSS stylesheet for each CSS rule.
-   *
-   * By default appends `<style>` element to `parent`.
-   *
-   * @param producer  Style producer instance.
-   *
-   * @returns CSS stylesheet reference.
-   */
-  addStyleSheet?: (producer: StyleProducer) => StyleSheetRef;
-
-  /**
-   * DOM rendering operations scheduler.
-   *
-   * Creates a render schedule per rule.
-   *
-   * Uses `newRenderSchedule` by default.
-   */
-  scheduler?: RenderScheduler;
-
-  /**
-   * Renderer or renderer chain to use.
-   */
-  renderer?: StypRenderer | readonly StypRenderer[];
-
-  /**
-   * Namespace aliaser to use.
-   *
-   * New instance will be created if not specified.
-   */
-  nsAlias?: NamespaceAliaser;
-
-}
-
-/**
- * CSS stylesheet reference.
- *
- * It is an object created by [[StypOptions.addStyleSheet]] option.
- *
- * @category Rendering
- */
-export interface StyleSheetRef {
-
-  /**
-   * CSS stylesheet reference.
-   */
-  readonly styleSheet: CSSStyleSheet;
-
-  /**
-   * Remoives stylesheet from the document.
-   */
-  remove(): void;
 
 }
